+++ src/binaries/hanami/src/api/http_endpoints/cluster/list_cluster_v1_0.rs
@@ -16,13 +16,14 @@ use actix_web::web::Json;
 use apistos::api_operation;
 
 use crate::config;
-use crate::database::host_table;
+use crate::database::meta_cluster_table;
 
 use ainari_api::common_functions::*;
 use ainari_api::errors::ErrorResponse;
 use ainari_api_structs::cluster_structs::*;
 use ainari_api_structs::user_context::UserContext;
-use ainari_clients::cluster as cluster_clients;
+use ainari_clients::endpoints::*;
+use ainari_clients::proxy as proxy_clients;
 
 #[api_operation(
     tag = "cluster",
@@ -32,26 +33,45 @@ use ainari_clients::cluster as cluster_clients;
     error_code = 500
 )]
 pub async fn list_cluster(context: UserContext) -> Result<Json<ClusterListResp>, ErrorResponse> {
-    let mut complete_resp = ClusterListResp {
+    // get clusters from db
+    let clusters = meta_cluster_table::list_meta_clusters(&context)
+        .map_err(|e| map_db_list_error("hosts", e))?;
+
+    // get endpoints from miko
+    let miko_endpoint = &config::CONFIG.miko;
+    let endpoints = get_endpoints(miko_endpoint, config::CONFIG.skip_tls_verification)
+        .await
+        .map_err(map_ainari_error_to_api_response)?;
+
+    // prepare response
+    let mut resp = ClusterListResp {
         clusters: Vec::new(),
     };
 
-    let host_list = host_table::list_hosts(&context).map_err(|e| map_db_list_error("hosts", e))?;
+    // fill reponse
+    for cluster in clusters {
+        let uuid = convert_uuid(&cluster.uuid)?;
 
-    for host in host_list {
-        let cluster_list_resp = cluster_clients::list_cluster(
-            &host.address,
+        // get port of the cluster from torii
+        let proxy_uuid = convert_uuid(&cluster.proxy_uuid)?;
+        let proxy_resp = proxy_clients::get_proxy(
+            &endpoints.torii,
             &context.token,
-            &config::INTERNAL_API_KEY,
+            &proxy_uuid,
             config::CONFIG.skip_tls_verification,
         )
         .await
         .map_err(map_ainari_error_to_api_response)?;
 
-        for cluster in cluster_list_resp.clusters {
-            complete_resp.clusters.push(cluster);
-        }
+        // add single object to the reponse-list
+        let obj = ClusterBasicResp {
+            uuid,
+            name: cluster.name.clone(),
+            proxy_port: proxy_resp.port,
+        };
+
+        resp.clusters.push(obj);
     }
 
-    Ok(Json(complete_resp))
+    Ok(Json(resp))
 }